feat(ui): implement header and footer components
