+++ packages/ui/src/components/header/index.tsx
@@ -1,0 +1,8 @@
+export function Header() {
+  return (
+    <div className="flex items-start m-6">
+      <h1 className="text-3xl font-bold text-zinc-800 dark:text-zinc-200">Uni Creator</h1>
+      <span className="ml-2 px-2 py-1 text-xs font-semibold bg-zinc-200 dark:bg-zinc-700 text-zinc-700 dark:text-zinc-300 rounded-lg">Beta</span>
+    </div>
+  )
+}
